Consolidate team commands into single module
- Merged `teamChat.js` and `teamHome.js` into `team.js` to reduce file clutter.
- Updated `index.js` to load the new `team.js` module.
- Updated `beforeChatSend.js` to import `toggleTeamChat` and `isTeamChatEnabled` from the new location.
- Deleted the obsolete `teamChat.js` and `teamHome.js` files.
- Ensured commands (`/teamchat`, `/hq`) and aliases (`/tc`, `/teamhome`) remain unchanged.
