+++ AddonExeBP/scripts/modules/commands/index.js
@@ -17,8 +17,7 @@ const commandFiles = [
 
     // --- TPA System ---
     'tpa.js',         // Contains /tpa, /tpahere, /tpaccept, /tpadeny, /tpacancel, /tpastatus
-    'teamChat.js',
-    'teamHome.js',
+    'team.js',        // Contains /teamchat, /hq
 
     // --- Home System ---
     'home.js',        // Contains /sethome, /delhome, /homes
